Automatically handle the Oban config in stage/prod env
