+++ lib/mix/tasks/database_safety.ex
@@ -12,8 +12,6 @@ defmodule Mix.Tasks.DatabaseSafety do
   use Mix.Task
   require Sanbase.Utils.Config, as: Config
 
-  @prod_db_patterns ["amazonaws"]
-
   @impl Mix.Task
   def run(_args) do
     if Code.ensure_loaded?(Envy) do
@@ -23,19 +21,8 @@ defmodule Mix.Tasks.DatabaseSafety do
     end
 
     env = Config.module_get(Sanbase, :env)
-    database_url = System.get_env("DATABASE_URL")
 
-    database_hostname = Config.module_get(Sanbase.Repo, :hostname)
-
-    prod_db_url? =
-      not is_nil(database_url) and
-        Enum.any?(@prod_db_patterns, &String.contains?(database_url, &1))
-
-    prod_db_config? =
-      not is_nil(database_hostname) and
-        Enum.any?(@prod_db_patterns, &String.contains?(database_hostname, &1))
-
-    case env != :prod and (prod_db_url? or prod_db_config?) do
+    case env != :prod and Sanbase.Utils.prod_db?() do
       true ->
         raise(Mix.Error, """
         Migration execution was stopped due to safety concerns!
